Squashed 'vendor/github.com/htop-dev/htop/' changes from 599233ffdf..4d170faae6
4d170faae6 PCP: Resolve nested derived metric definitions for devices
8dfe7ed645 Improve `make dist` to warn on building inside a git repository
a94e7661ba Merge branch 'uaf' of cgzones/htop
66f0ba0478 Linux: reorder /proc/<pid>/status parsing
314d693c4b Annotate mbstowcs_nonfatal() with restrict
ef07d6ed9f Linux: show newlines in commands as ? instead of space
4b3dfa246e CI: Bump NetBSD version to 10.0
d2ec14fa4a Fix GCC build warning in NetBSD 10
e54871b679 Linux: Permit non-ASCII for CGROUP, CONTAINER & SECATTR fields
c63dc63a71 Fix "CPU usage bar" help alignment.
30618c586c Update '__STDC_VERSION__' check with C23
4c0e96592d darwin: scan thread information
0c002751dd darwin: avoid double division by zero on startup
6f142695cb Use struct member to determine size of allocations
784ef0878a Adjust some code style for maintainability
e0eed79da6 MemorySwapMeter_init() code shrink
c2c2a70997 Define MeterModeId to unsigned int and use it throughout
c167a2c4c5 New header "MeterMode.h" for MeterModeId definitions
24b5491251 Make 'Meter_modes' private and simplify its structure
3e00de8e68 Reorder Meter.c sections. No code changes.
74d83d82d2 MemorySwapMeter: Read height from sub-meter objects
c711dbc633 CPUMeter: Read height from sub-meter objects
2fa283d202 Remove CUSTOM_METERMODE MeterModeId
f795e84085 Don't initialize meter mode in custom meter init functions
faaf6dd25e Obsolete the CUSTOM_METERMODE use in default mode
52486dbc0f Merge branch 'gcc14-analyzer-fixes' of BenBE/htop
0a1db8a770 Avoid magic numbers for the size of FunctionBar lists
7017b86bbf Use 'sb' name for local 'struct stat' buffers
a46b3f0a98 Use 'fp' name for local 'FILE*' variables.
9cc1c5ea15 Reserve full function array to make the GCC14 static code analyzer happy
51ead9d8b6 Explicit memory initialization when reading status file
6f0adfab24 Merge branch 'add-git-commit-to-compiled-dev-versions' of fasterit/htop
6eb164ef68 Merge branch 'hf-one-100' of peterdey/htop
e8ee2004bf Merge branch 'signal_safe' of cgzones/htop
8ad2bb8e4f Merge branch 'htoprc-write-check' of Explorer09/htop-1
13ca4e2b5f Add HeaderLayout format for one single 100%-width column
752659fd08 Avoid writing configuration to non-regular files
eaeeda64c1 Memory leak when a legacy config file exists
2acd62d3af Update configure.ac with some m4 magic to add a git commit id  and a reference of the last tag, the changes since then and the  dirty flag from `git describe`
8122fc394e CI: (OpenBSD) Detect latest Autoconf & Automake versions
550f101d7e CI: Remove Autoconf version num (FreeBSD & DragonFlyBSD)
81ff634427 Avoid fprintf in Settings_write
2fafa349a7 Avoid fprintf in signal handlers
49bad52fc2 linux: harden terminal parsing
0da63c903c BarMeterMode_draw() code formatting adjustment
1a12d58526 Work around GCC14 memleak diagnostic
327593e5cd Align parameter names in declarations with definitions
4cf8c98e05 Drop return at end of function returning void
56d7622902 Use uppercase floating point literals
626db518b3 Drop casts to same type
3477fbf2c9 Linux: fix title alignments of GPU columns
0b0d50c50e Row_printNanoseconds fixes

git-subtree-dir: vendor/github.com/htop-dev/htop
git-subtree-split: 4d170faae6aff29619b2d78955182761ce467a7a

+++ Settings.c
@@ -11,6 +11,7 @@ in the source distribution for its full text.
 
 #include <ctype.h>
 #include <errno.h>
+#include <fcntl.h>
 #include <limits.h>
 #include <pwd.h>
 #include <stdio.h>
@@ -61,13 +62,13 @@ static char** readQuotedList(char* line) {
    return list;
 }
 
-static void writeQuotedList(FILE* fd, char** list) {
+static void writeQuotedList(FILE* fp, char** list) {
    const char* sep = "";
    for (int i = 0; list[i]; i++) {
-      fprintf(fd, "%s\"%s\"", sep, list[i]);
+      fprintf(fp, "%s\"%s\"", sep, list[i]);
       sep = " ";
    }
-   fprintf(fd, "\n");
+   fprintf(fp, "\n");
 }
 
 */
@@ -107,9 +108,9 @@ static void Settings_readMeterModes(Settings* this, const char* line, unsigned i
    }
    column = MINIMUM(column, HeaderLayout_getColumns(this->hLayout) - 1);
    this->hColumns[column].len = len;
-   int* modes = len ? xCalloc(len, sizeof(int)) : NULL;
+   MeterModeId* modes = len ? xCalloc(len, sizeof(MeterModeId)) : NULL;
    for (int i = 0; i < len; i++) {
-      modes[i] = atoi(ids[i]);
+      modes[i] = (MeterModeId) atoi(ids[i]);
    }
    String_freeArray(ids);
    this->hColumns[column].modes = modes;
@@ -122,7 +123,7 @@ static bool Settings_validateMeters(Settings* this) {
 
    for (size_t column = 0; column < colCount; column++) {
       char** names = this->hColumns[column].names;
-      const int* modes = this->hColumns[column].modes;
+      const MeterModeId* modes = this->hColumns[column].modes;
       const size_t len = this->hColumns[column].len;
 
       if (!len)
@@ -162,8 +163,8 @@ static void Settings_defaultMeters(Settings* this, unsigned int initialCpuCount)
    this->hLayout = HF_TWO_50_50;
    this->hColumns = xCalloc(HeaderLayout_getColumns(this->hLayout), sizeof(MeterColumnSetting));
    for (size_t i = 0; i < 2; i++) {
-      this->hColumns[i].names = xCalloc(sizes[i] + 1, sizeof(char*));
-      this->hColumns[i].modes = xCalloc(sizes[i], sizeof(int));
+      this->hColumns[i].names = xCalloc(sizes[i] + 1, sizeof(*this->hColumns[0].names));
+      this->hColumns[i].modes = xCalloc(sizes[i], sizeof(*this->hColumns[0].modes));
       this->hColumns[i].len = sizes[i];
    }
 
@@ -232,7 +233,7 @@ static const char* toFieldName(Hashtable* columns, int id, bool* enabled) {
 }
 
 static int toFieldIndex(Hashtable* columns, const char* str) {
-   if (isdigit(str[0])) {
+   if (isdigit((unsigned char)str[0])) {
       // This "+1" is for compatibility with the older enum format.
       int id = atoi(str) + 1;
       if (toFieldName(columns, id, NULL)) {
@@ -353,16 +354,50 @@ static ScreenSettings* Settings_defaultScreens(Settings* this) {
    return this->screens[0];
 }
 
-static bool Settings_read(Settings* this, const char* fileName, unsigned int initialCpuCount) {
-   FILE* fd = fopen(fileName, "r");
-   if (!fd)
+static bool Settings_read(Settings* this, const char* fileName, unsigned int initialCpuCount, bool checkWritability) {
+   int fd = -1;
+   const char* fopen_mode = "r+";
+   if (checkWritability) {
+      do {
+         fd = open(fileName, O_RDWR | O_NOCTTY | O_NOFOLLOW);
+      } while (fd < 0 && errno == EINTR);
+
+      if (fd < 0) {
+         this->writeConfig = (errno == ENOENT);
+         if (errno != EACCES && errno != EPERM && errno != EROFS) {
+            return false;
+         }
+      } else {
+         // Check if this is a regular file
+         struct stat sb;
+         int err = fstat(fd, &sb);
+         this->writeConfig = !err && S_ISREG(sb.st_mode);
+      }
+   }
+
+   // If opening for read & write is not needed or fails, open for read only.
+   // There is no risk of following symlink in this case.
+   if (fd < 0) {
+      fopen_mode = "r";
+      do {
+         fd = open(fileName, O_RDONLY | O_NOCTTY);
+      } while (fd < 0 && errno == EINTR);
+   }
+
+   if (fd < 0)
       return false;
+
+   FILE* fp = fdopen(fd, fopen_mode);
+   if (!fp) {
+      close(fd);
+      return false;
+   }
 
    ScreenSettings* screen = NULL;
    bool didReadMeters = false;
    bool didReadAny = false;
    for (;;) {
-      char* line = String_readLine(fd);
+      char* line = String_readLine(fp);
       if (!line) {
          break;
       }
@@ -382,7 +417,7 @@ static bool Settings_read(Settings* this, const char* fileName, unsigned int ini
             fprintf(stderr, "         version v%d, but this %s binary only supports up to version v%d.\n", this->config_version, PACKAGE, CONFIG_READER_MIN_VERSION);
             fprintf(stderr, "         The configuration file will be downgraded to v%d when %s exits.\n", CONFIG_READER_MIN_VERSION, PACKAGE);
             String_freeArray(option);
-            fclose(fd);
+            fclose(fp);
             return false;
          }
       } else if (String_eq(option[0], "fields") && this->config_version <= 2) {
@@ -555,7 +590,7 @@ static bool Settings_read(Settings* this, const char* fileName, unsigned int ini
       }
       String_freeArray(option);
    }
-   fclose(fd);
+   fclose(fp);
    if (!didReadMeters || !Settings_validateMeters(this))
       Settings_defaultMeters(this, initialCpuCount);
    if (!this->nScreens)
@@ -563,65 +598,90 @@ static bool Settings_read(Settings* this, const char* fileName, unsigned int ini
    return didReadAny;
 }
 
-static void writeFields(FILE* fd, const ProcessField* fields, Hashtable* columns, bool byName, char separator) {
+typedef ATTR_FORMAT(printf, 2, 3) int (*OutputFunc)(FILE*, const char*,...);
+
+static void writeFields(OutputFunc of, FILE* fp,
+                        const ProcessField* fields, Hashtable* columns,
+                        bool byName, char separator) {
    const char* sep = "";
    for (unsigned int i = 0; fields[i]; i++) {
       if (fields[i] < LAST_PROCESSFIELD && byName) {
          const char* pName = toFieldName(columns, fields[i], NULL);
-         fprintf(fd, "%s%s", sep, pName);
+         of(fp, "%s%s", sep, pName);
       } else if (fields[i] >= LAST_PROCESSFIELD && byName) {
          bool enabled;
          const char* pName = toFieldName(columns, fields[i], &enabled);
          if (enabled)
-            fprintf(fd, "%sDynamic(%s)", sep, pName);
+            of(fp, "%sDynamic(%s)", sep, pName);
       } else {
          // This "-1" is for compatibility with the older enum format.
-         fprintf(fd, "%s%d", sep, (int) fields[i] - 1);
+         of(fp, "%s%d", sep, (int) fields[i] - 1);
       }
       sep = " ";
    }
-   fputc(separator, fd);
+   of(fp, "%c", separator);
 }
 
-static void writeList(FILE* fd, char** list, int len, char separator) {
+static void writeList(OutputFunc of, FILE* fp,
+                      char** list, int len, char separator) {
    const char* sep = "";
    for (int i = 0; i < len; i++) {
-      fprintf(fd, "%s%s", sep, list[i]);
+      of(fp, "%s%s", sep, list[i]);
       sep = " ";
    }
-   fputc(separator, fd);
+   of(fp, "%c", separator);
 }
 
-static void writeMeters(const Settings* this, FILE* fd, char separator, unsigned int column) {
+static void writeMeters(const Settings* this, OutputFunc of,
+                        FILE* fp, char separator, unsigned int column) {
    if (this->hColumns[column].len) {
-      writeList(fd, this->hColumns[column].names, this->hColumns[column].len, separator);
+      writeList(of, fp, this->hColumns[column].names, this->hColumns[column].len, separator);
    } else {
-      fputc('!', fd);
-      fputc(separator, fd);
+      of(fp, "!%c", separator);
    }
 }
 
-static void writeMeterModes(const Settings* this, FILE* fd, char separator, unsigned int column) {
+static void writeMeterModes(const Settings* this, OutputFunc of,
+                            FILE* fp, char separator, unsigned int column) {
    if (this->hColumns[column].len) {
       const char* sep = "";
       for (size_t i = 0; i < this->hColumns[column].len; i++) {
-         fprintf(fd, "%s%d", sep, this->hColumns[column].modes[i]);
+         of(fp, "%s%u", sep, this->hColumns[column].modes[i]);
          sep = " ";
       }
    } else {
-      fputc('!', fd);
+      of(fp, "!");
    }
 
-   fputc(separator, fd);
+   of(fp, "%c", separator);
+}
+
+ATTR_FORMAT(printf, 2, 3)
+static int signal_safe_fprintf(FILE* stream, const char* fmt, ...) {
+   char buf[2048];
+
+   va_list vl;
+   va_start(vl, fmt);
+   int n = vsnprintf(buf, sizeof(buf), fmt, vl);
+   va_end(vl);
+
+   if (n <= 0)
+      return n;
+
+   return full_write_str(fileno(stream), buf);
 }
 
 int Settings_write(const Settings* this, bool onCrash) {
-   FILE* fd;
+   FILE* fp;
    char separator;
    char* tmpFilename = NULL;
+   OutputFunc of;
    if (onCrash) {
-      fd = stderr;
+      fp = stderr;
       separator = ';';
+      of = signal_safe_fprintf;
+   } else if (!this->writeConfig) {
+      return 0;
    } else {
       /* create tempfile with mode 0600 */
       mode_t cur_umask = umask(S_IXUSR | S_IRWXG | S_IRWXO);
@@ -630,24 +690,25 @@ int Settings_write(const Settings* this, bool onCrash) {
       umask(cur_umask);
       if (fdtmp == -1)
          return -errno;
-      fd = fdopen(fdtmp, "w");
-      if (fd == NULL)
+      fp = fdopen(fdtmp, "w");
+      if (!fp)
          return -errno;
       separator = '\n';
+      of = fprintf;
    }
 
    #define printSettingInteger(setting_, value_) \
-      fprintf(fd, setting_ "=%d%c", (int) (value_), separator)
+      of(fp, setting_ "=%d%c", (int) (value_), separator)
    #define printSettingString(setting_, value_) \
-      fprintf(fd, setting_ "=%s%c", value_, separator)
+      of(fp, setting_ "=%s%c", value_, separator)
 
    if (!onCrash) {
-      fprintf(fd, "# Beware! This file is rewritten by htop when settings are changed in the interface.\n");
-      fprintf(fd, "# The parser is also very primitive, and not human-friendly.\n");
+      of(fp, "# Beware! This file is rewritten by htop when settings are changed in the interface.\n");
+      of(fp, "# The parser is also very primitive, and not human-friendly.\n");
    }
    printSettingString("htop_version", VERSION);
    printSettingInteger("config_reader_min_version", CONFIG_READER_MIN_VERSION);
-   fprintf(fd, "fields="); writeFields(fd, this->screens[0]->fields, this->dynamicColumns, false, separator);
+   of(fp, "fields="); writeFields(of, fp, this->screens[0]->fields, this->dynamicColumns, false, separator);
    printSettingInteger("hide_kernel_threads", this->hideKernelThreads);
    printSettingInteger("hide_userland_threads", this->hideUserlandThreads);
    printSettingInteger("hide_running_in_container", this->hideRunningInContainer);
@@ -688,10 +749,10 @@ int Settings_write(const Settings* this, bool onCrash) {
 
    printSettingString("header_layout", HeaderLayout_getName(this->hLayout));
    for (unsigned int i = 0; i < HeaderLayout_getColumns(this->hLayout); i++) {
-      fprintf(fd, "column_meters_%u=", i);
-      writeMeters(this, fd, separator, i);
-      fprintf(fd, "column_meter_modes_%u=", i);
-      writeMeterModes(this, fd, separator, i);
+      of(fp, "column_meters_%u=", i);
+      writeMeters(this, of, fp, separator, i);
+      of(fp, "column_meter_modes_%u=", i);
+      writeMeterModes(this, of, fp, separator, i);
    }
 
    // Legacy compatibility with older versions of htop
@@ -709,14 +770,14 @@ int Settings_write(const Settings* this, bool onCrash) {
       const char* sortKey = toFieldName(this->dynamicColumns, ss->sortKey, NULL);
       const char* treeSortKey = toFieldName(this->dynamicColumns, ss->treeSortKey, NULL);
 
-      fprintf(fd, "screen:%s=", ss->heading);
-      writeFields(fd, ss->fields, this->dynamicColumns, true, separator);
+      of(fp, "screen:%s=", ss->heading);
+      writeFields(of, fp, ss->fields, this->dynamicColumns, true, separator);
       if (ss->dynamic) {
          printSettingString(".dynamic", ss->dynamic);
          if (ss->sortKey && ss->sortKey != PID)
-            fprintf(fd, "%s=Dynamic(%s)%c", ".sort_key", sortKey, separator);
+            of(fp, "%s=Dynamic(%s)%c", ".sort_key", sortKey, separator);
          if (ss->treeSortKey && ss->treeSortKey != PID)
-            fprintf(fd, "%s=Dynamic(%s)%c", ".tree_sort_key", treeSortKey, separator);
+            of(fp, "%s=Dynamic(%s)%c", ".tree_sort_key", treeSortKey, separator);
       } else {
          printSettingString(".sort_key", sortKey);
          printSettingString(".tree_sort_key", treeSortKey);
@@ -736,10 +797,10 @@ int Settings_write(const Settings* this, bool onCrash) {
 
    int r = 0;
 
-   if (ferror(fd) != 0)
+   if (ferror(fp) != 0)
       r = (errno != 0) ? -errno : -EBADF;
 
-   if (fclose(fd) != 0)
+   if (fclose(fp) != 0)
       r = r ? r : -errno;
 
    if (r == 0)
@@ -752,6 +813,8 @@ int Settings_write(const Settings* this, bool onCrash) {
 
 Settings* Settings_new(unsigned int initialCpuCount, Hashtable* dynamicMeters, Hashtable* dynamicColumns, Hashtable* dynamicScreens) {
    Settings* this = xCalloc(1, sizeof(Settings));
+
+   this->writeConfig = true;
 
    this->dynamicScreens = dynamicScreens;
    this->dynamicColumns = dynamicColumns;
@@ -815,17 +878,12 @@ Settings* Settings_new(unsigned int initialCpuCount, Hashtable* dynamicMeters, H
          configDir = String_cat(home, "/.config");
          htopDir = String_cat(home, "/.config/htop");
       }
-      legacyDotfile = String_cat(home, "/.htoprc");
       (void) mkdir(configDir, 0700);
       (void) mkdir(htopDir, 0700);
       free(htopDir);
       free(configDir);
-      struct stat st;
-      int err = lstat(legacyDotfile, &st);
-      if (err || S_ISLNK(st.st_mode)) {
-         free(legacyDotfile);
-         legacyDotfile = NULL;
-      }
+
+      legacyDotfile = String_cat(home, "/.htoprc");
    }
 
    this->filename = xMalloc(PATH_MAX);
@@ -838,21 +896,22 @@ Settings* Settings_new(unsigned int initialCpuCount, Hashtable* dynamicMeters, H
 #endif
    this->changed = false;
    this->delay = DEFAULT_DELAY;
-   bool ok = Settings_read(this, this->filename, initialCpuCount);
+
+   bool ok = Settings_read(this, this->filename, initialCpuCount, /*checkWritability*/true);
    if (!ok && legacyDotfile) {
-      ok = Settings_read(this, legacyDotfile, initialCpuCount);
-      if (ok) {
+      ok = Settings_read(this, legacyDotfile, initialCpuCount, this->writeConfig);
+      if (ok && this->writeConfig) {
          // Transition to new location and delete old configuration file
          if (Settings_write(this, false) == 0) {
             unlink(legacyDotfile);
          }
       }
-      free(legacyDotfile);
    }
    if (!ok) {
       this->screenTabs = true;
       this->changed = true;
-      ok = Settings_read(this, SYSCONFDIR "/htoprc", initialCpuCount);
+
+      ok = Settings_read(this, SYSCONFDIR "/htoprc", initialCpuCount, /*checkWritability*/false);
    }
    if (!ok) {
       Settings_defaultMeters(this, initialCpuCount);
@@ -863,6 +922,8 @@ Settings* Settings_new(unsigned int initialCpuCount, Hashtable* dynamicMeters, H
    this->ss = this->screens[this->ssIndex];
 
    this->lastUpdate = 1;
+
+   free(legacyDotfile);
 
    return this;
 }
